pyproject.toml
[project]
name = "django-playground"
version = "0.1.0"
description = "Happy fun play time"
dependencies = [
    "Django>=5.0.2",
    "python-dotenv>=1.0.1",
]
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/django_playground"]
